// root/src/modules/interfaces/enums/search/FieldType.ts
export enum FieldType {
    None = "none",
    All = "all",
    Name = "name",
    Description = 'description',
    Tags = 'tags',
    Owner = 'owner',
    IsStash = 'isStash',
    IsFlat = 'isFlat',
    IsLoop = 'isLoop',
    TreeToken = 'treeToken',
    IsSubtree = 'isSubtree',
    IsPublished = 'isPublished',
    Discussion = 'discussion',
    Option = 'option',
    IsEntry = 'isEntry',
    IsSocket = 'isSocket',
    IsRoot = 'isRoot',
    IsLink = 'isLink',
    IsPlug = 'isPlug',
    IsParentRoot = 'isParentRoot',
    IsSolution = 'isSolution',
    Order = 'order'
}